Restructuring of `Parser` exception hierarchy
